[CLEANUP] Consolidate Scope classes
There are two scope interfaces that exist currently, and two
implementations, but they exist in separate files. The environment file
is already over croweded, so this PR moves the PartialScope to the
same file as dynamic scope, and renames the classes so they are more
consistent overall.
