+++ packages/@glimmer/runtime/lib/compiled/opcodes/-debug-strip.ts
@@ -28,7 +28,7 @@ import {
 } from '@glimmer/interfaces';
 import { PathReference, Reference } from '@glimmer/reference';
 import { Tag, COMPUTE } from '@glimmer/validator';
-import { ScopeImpl } from '../../environment';
+import { PartialScopeImpl } from '../../scope';
 import CurryComponentReference from '../../references/curry-component';
 import {
   CapturedNamedArgumentsImpl,
@@ -96,7 +96,9 @@ export const CheckCapturedArguments: Checker<CapturedArguments> = CheckInterface
 
 export const CheckCurryComponent = wrap(() => CheckInstanceof(CurryComponentReference));
 
-export const CheckScope: Checker<Scope<JitOrAotBlock>> = wrap(() => CheckInstanceof(ScopeImpl));
+export const CheckScope: Checker<Scope<JitOrAotBlock>> = wrap(() =>
+  CheckInstanceof(PartialScopeImpl)
+);
 
 export const CheckComponentManager: Checker<ComponentManager<unknown>> = CheckInterface({
   getCapabilities: CheckFunction,
