+++ packages/@glimmer/runtime/lib/opcodes.ts
@@ -10,7 +10,7 @@ import { debug, logOpcode } from '@glimmer/debug';
 import { DESTROYABLE_STACK, INNER_VM, CONSTANTS, STACKS } from './symbols';
 import { InternalVM, InternalJitVM } from './vm/append';
 import { CURSOR_STACK } from './vm/element-builder';
-import { isScopeReference } from './environment';
+import { isScopeReference } from './scope';
 
 export interface OpcodeJSON {
   type: number | string;
